Remove test-utils helper crate
Benefits of doing this:

* We don't have to worry about the rust-analyzer bug (which bugs me
  quite a bit because "go to defintion" does not work, for example)

* One less crate to worry about. And the fewer crates, the simpler =
  better.

* The "assert stdout or update stdout" paths becomes less annoying.

Downsides:

* This leads to some code duplication, but it's pretty mild.
