--- test-utils/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "test-utils"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-
-
-[dependencies]
-anyhow = "1.0.53"
-assert_cmd = "2.0.4"
-expect-test = "1.4.0"
-pretty_assertions = "1.3.0"
-tempfile = "3.3.0"
-
-[dependencies.rustdoc-json]
-path = "../rustdoc-json"
-version = "*"
